Enable LAVA Apache configuration
HTTP port from guest machine is forwarded to the host.

Port forwarding allows accessing specific port within the guest machine
from a port on the host machine. In this case, accessing
"localhost:8080" will access port 80 on the guest machine.

LAVA will be the main service hosted on this virtual machine.